+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v0.0.3 (2025-12-29)
+
+- **Summary:** Update 7 files
+- **Files:**
+  - `.githooks/pre-commit`
+  - `.githooks/prepare-commit-msg`
+  - `CMakeLists.txt`
+  - `Source/PluginEditor.cpp`
+  - `Source/PluginEditor.h`
+  - `Source/PluginProcessor.cpp`
+  - `Source/PluginProcessor.h`
+
 ## v0.0.2 (2025-12-29)
 
 - **Summary:** Update 2 files